Crate comlib_geometry

Comlib Geometric Primitives and Algorithms

This crate provides geometric primitives and algorithms that work on them. The main primitive types are Point and Line. For convenience the crate also provides Segment and Polygon types.

Currently only the following algorithms have been implemented:

Convex hull

Structs

Line

Represents the set of points (x, y) which satisfy ax + by + c = 0.

Point

Polygon

PolygonSegmentIter

Segment

Unit

Enums

LineIntersection

Ordering

SegmentIntersection

Traits

ValidCoordinate

Functions

convex_hull

Computes the convex hull of the given set of points.